simplify create_modules() 'route' option
Thanks a million for your blogs! Im currently going through the entire setup and loving it :D
I thought I'd give back a bit by (starting with) this simplification.
Correct me if I'm wrong, but I think all you are trying to achieve is to sum the number of output filters for the 1 or 2 layers a route-layer is referring to. This can be done in 1 line as proposed.

I'm not a 100% sure if Redmon is using 0 indexing when referring to his layers, but I believe he is.